mv userName to parent (messageDisplay) to make it accessible in messageBox
